+++ TeamManagementApp/Views/Members/IMembersView.cs
@@ -10,10 +10,10 @@ namespace TeamManagementApp.Views.Members
 {
     public interface IMembersView
     {
-        ICommonComboBox<Member, int> CmbMembers { get; set; }
+        ICommonComboBox<Member, long> CmbMembers { get; set; }
         ICommonTextBox TxtMemberDetails { get; set; }
-        ICommonComboBox<MemberClass, int> CmbClass { get; set; }
-        ICommonComboBox<UserLogin, int> CmbUser { get; set; }
+        ICommonComboBox<MemberClass, long> CmbClass { get; set; }
+        ICommonComboBox<UserLogin, long> CmbUser { get; set; }
         ICommonCheckBox ChkActive { get; set; }
     }
 }
